Move Cell class to its own jsx
